usbForensicsTool/requirements.txt
pyusb
pyudev
libusb1
volatility3
pandas